refactor CPU-bound SDF API
