+++ go.mod
@@ -6,7 +6,9 @@ require (
 	github.com/chewxy/math32 v1.10.1
 	github.com/deadsy/sdfx v0.0.0-20220428051248-ab3af168a1af
 	github.com/fogleman/fauxgl v0.0.0-20200818143847-27cddc103802
+	github.com/go-gl/gl v0.0.0-20211210172815-726fda9656d6
 	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646
+	github.com/soypat/glgl v0.0.0-20240727212338-00ef53124943
 	gonum.org/v1/gonum v0.11.1-0.20220625074215-67f3e1dbfccc
 	gonum.org/v1/plot v0.11.0
 )
@@ -15,11 +17,13 @@ require (
 	github.com/ajstarks/svgo v0.0.0-20211024235047-1546f124cd8b // indirect
 	github.com/dhconnelly/rtreego v1.1.0 // indirect
 	github.com/fogleman/simplify v0.0.0-20170216171241-d32f302d5046 // indirect
+	github.com/go-gl/glfw v0.0.0-20221017161538-93cebf72946b // indirect
+	github.com/go-gl/glfw/v3.3/glfw v0.0.0-20221017161538-93cebf72946b // indirect
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
 	github.com/hschendel/stl v1.0.4 // indirect
 	github.com/llgcode/draw2d v0.0.0-20200930101115-bfaf5d914d1e // indirect
 	github.com/yofu/dxf v0.0.0-20190710012328-5a6d1e83f16c // indirect
-	golang.org/x/exp v0.0.0-20220613132600-b0d781184e0d // indirect
+	golang.org/x/exp v0.0.0-20221230185412-738e83a70c30 // indirect
 	golang.org/x/image v0.0.0-20220617043117-41969df76e82 // indirect
 	rsc.io/pdf v0.1.1 // indirect
 )
